feat: 비밀번호 유효성 검사

+++ controllers/ldap.py
@@ -190,16 +190,22 @@ def delete_group(user_id):
            response = e
     return response
 
-def delete_user(user_id):
+def delete_user(user_id, password):
     ldap_conn = connect_ldap_server()
     user_dn = "cn="+user_id+",ou=users,cn=admin,dc=devstack,dc=co,dc=kr"
-
+    
+    # 사용자 존재 여부 확인
     check = check_user(user_id)
     if len(check) == 0:
         return "user"
 
+    # 비밀번호 검증
+    check = check_password_validation(user_id, password)
+    if check == False:
+        return "passwd"
+
+    ks_project.delete_project(user_id)
     delete_group(user_id)
-    #ks_project.delete_project(user_name)
     ks_user.delete_user(user_id)
 
     try:
@@ -207,3 +213,16 @@ def delete_user(user_id):
     except LDAPException as e:
         response = e
     return response
+
+
+def check_password_validation(user_id, password):
+    try:
+        server_uri = config.LDAP_URL
+        server = Server(server_uri, get_info=ALL)
+        connection = Connection(server,
+                        user='cn='+user_id+',ou=users,'+config.LDAP_ADMIN_DN,
+                        password=password)
+        bind_response = connection.bind()
+    except LDAPBindError as e:
+        bind_response = e
+    return bind_response
